Expanded on the previous commit for Issue #10
mostly just wrote startup code that has not yet been tested.  Wrote a
function that will get called every 3 hours to check to see if the log
file has exceeded 1MB, if so, it creates a new log file.  Also created a
function that gets called every 5 seconds to write some information to
the log file (which I haven't decided yet).  Lastly, I expanded onto the
'/readLogFile' request so that it will return all the data contained in
the current Log file.